Moved schedule functions to utils
