+++ histomicstk/__init__.py
@@ -6,21 +6,12 @@ import preprocessing
 import segmentation
 import utils
 
-from .ConvertSchedule import ConvertSchedule
-from .Sample import Sample
-from .SubmitTorque import SubmitTorque
-from .TilingSchedule import TilingSchedule
-
 __version__ = '0.1.0'
 
 __all__ = (
-    'ConvertSchedule',
     'features',
     'filters',
     'preprocessing',
-    'Sample',
     'segmentation',
-    'SubmitTorque',
-    'TilingSchedule',
     'utils',
 )
