+++ histomicstk/preprocessing/color_normalization/ReinhardSample.py
@@ -1,6 +1,6 @@
 import collections
 import numpy as np
-from histomicstk import Sample
+from histomicstk.utils import Sample
 from histomicstk.preprocessing.color_conversion import RudermanLABFwd
