+++ histomicstk/utils/__init__.py
@@ -1,15 +1,23 @@
+from .ConvertSchedule import ConvertSchedule
 from .Del2 import Del2
 from .Eigenvalues import Eigenvalues
 from .GradientDiffusion import GradientDiffusion
 from .Hessian import Hessian
 from .MergeColinear import MergeColinear
 from .PoissonMixture import PoissonMixture
+from .Sample import Sample
+from .SubmitTorque import SubmitTorque
+from .TilingSchedule import TilingSchedule
 
 __all__ = (
+    'ConvertSchedule',
     'Del2',
     'Eigenvalues',
     'GradientDiffusion',
     'Hessian',
     'MergeColinear',
     'PoissonMixture',
+    'Sample',
+    'SubmitTorque',
+    'TilingScheule',
 )
